SCRUM-78 current logged in user can't be deleted from the UI.

+++ stimmungsumfrage/src/main/java/org/lf10/stimmungsumfrage/Controllers/UserController.java
@@ -2,6 +2,9 @@ package org.lf10.stimmungsumfrage.Controllers;
 
 import lombok.RequiredArgsConstructor;
 import org.lf10.stimmungsumfrage.Models.Forms.FeedbackForm;
+
+import java.io.Console;
+
 import org.lf10.stimmungsumfrage.Models.User;
 import org.lf10.stimmungsumfrage.Repositories.DepartmentRepository;
 import org.lf10.stimmungsumfrage.Repositories.RoleRepository;
@@ -12,6 +15,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -77,7 +81,11 @@ public class UserController {
     @DeleteMapping("/{id}")  // Use POST for final delete (safer)
     public String deleteUser(@PathVariable Long id) {
         User user = userRepository.findById(id).get();
-        userRepository.delete(user);
+        String currentUsersEmail = SecurityContextHolder.getContext().getAuthentication().getName();
+        if (!currentUsersEmail.equals(user.getEmail()))
+        {
+            userRepository.delete(user);
+        }
         return "redirect:/admin/users";
     }
